+++ main.go
@@ -46,6 +46,7 @@ import (
 	commonmonitoring "github.com/giantswarm/observability-operator/pkg/common/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
 	"github.com/giantswarm/observability-operator/pkg/common/password"
+	"github.com/giantswarm/observability-operator/pkg/grafana/client"
 	"github.com/giantswarm/observability-operator/pkg/monitoring"
 	"github.com/giantswarm/observability-operator/pkg/monitoring/alloy"
 	"github.com/giantswarm/observability-operator/pkg/monitoring/heartbeat"
@@ -81,7 +82,12 @@ var (
 )
 
 const (
-	OpsgenieApiKey = "OPSGENIE_API_KEY" // #nosec G101
+	grafanaAdminUsernameEnvVar = "GRAFANA_ADMIN_USERNAME" // #nosec G101
+	grafanaAdminPasswordEnvVar = "GRAFANA_ADMIN_PASSWORD" // #nosec G101
+	grafanaTLSCertFileEnvVar   = "GRAFANA_TLS_CERT_FILE"  // #nosec G101
+	grafanaTLSKeyFileEnvVar    = "GRAFANA_TLS_KEY_FILE"   // #nosec G101
+
+	opsgenieApiKeyEnvVar = "OPSGENIE_API_KEY" // #nosec G101
 )
 
 func init() {
@@ -117,10 +123,10 @@ func main() {
 	flag.StringVar(&managementClusterRegion, "management-cluster-region", "",
 		"The region of the management cluster.")
 	// Monitoring configuration flags.
+	flag.StringVar(&monitoringAgent, "monitoring-agent", commonmonitoring.MonitoringAgentAlloy,
+		fmt.Sprintf("select monitoring agent to use (%s or %s)", commonmonitoring.MonitoringAgentPrometheus, commonmonitoring.MonitoringAgentAlloy))
 	flag.BoolVar(&monitoringEnabled, "monitoring-enabled", false,
 		"Enable monitoring at the management cluster level.")
-	flag.StringVar(&monitoringAgent, "monitoring-agent", commonmonitoring.MonitoringAgentPrometheus,
-		fmt.Sprintf("select monitoring agent to use (%s or %s)", commonmonitoring.MonitoringAgentPrometheus, commonmonitoring.MonitoringAgentAlloy))
 	flag.Float64Var(&monitoringShardingScaleUpSeriesCount, "monitoring-sharding-scale-up-series-count", 0,
 		"Configures the number of time series needed to add an extra prometheus agent shard.")
 	flag.Float64Var(&monitoringShardingScaleDownPercentage, "monitoring-sharding-scale-down-percentage", 0,
@@ -197,9 +203,9 @@ func main() {
 		Region:     managementClusterRegion,
 	}
 
-	var opsgenieApiKey = os.Getenv(OpsgenieApiKey)
+	var opsgenieApiKey = os.Getenv(opsgenieApiKeyEnvVar)
 	if opsgenieApiKey == "" {
-		setupLog.Error(nil, fmt.Sprintf("environment variable %s not set", OpsgenieApiKey))
+		setupLog.Error(nil, fmt.Sprintf("environment variable %s not set", opsgenieApiKeyEnvVar))
 		os.Exit(1)
 	}
 
@@ -257,9 +263,35 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "Cluster")
 		os.Exit(1)
 	}
+
+	// Generate Grafana client
+	// Get grafana admin-password and admin-user
+	grafanaAdminCredentials := client.AdminCredentials{
+		Username: os.Getenv(grafanaAdminUsernameEnvVar),
+		Password: os.Getenv(grafanaAdminPasswordEnvVar),
+	}
+	if grafanaAdminCredentials.Username == "" {
+		setupLog.Error(nil, fmt.Sprintf("environment variable %s not set", grafanaAdminUsernameEnvVar))
+		os.Exit(1)
+	}
+	if grafanaAdminCredentials.Password == "" {
+		setupLog.Error(nil, fmt.Sprintf("environment variable %s not set", grafanaAdminPasswordEnvVar))
+		os.Exit(1)
+	}
+	grafanaTLSConfig := client.TLSConfig{
+		Cert: os.Getenv(grafanaTLSCertFileEnvVar),
+		Key:  os.Getenv(grafanaTLSKeyFileEnvVar),
+	}
+	grafanaAPI, err := client.GenerateGrafanaClient(grafanaAdminCredentials, grafanaTLSConfig)
+	if err != nil {
+		setupLog.Error(err, "unable to create grafana client")
+		os.Exit(1)
+	}
+
 	if err = (&controller.GrafanaOrganizationReconciler{
-		Client: mgr.GetClient(),
-		Scheme: mgr.GetScheme(),
+		Client:     mgr.GetClient(),
+		Scheme:     mgr.GetScheme(),
+		GrafanaAPI: grafanaAPI,
 	}).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "GrafanaOrganization")
 		os.Exit(1)