+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2024-10-17
+
+### Fixed
+
+- Fix `flag redefined` error
+
+## [0.8.0] - 2024-10-17
+
 ### Added
 
 - Add wal `truncate_frequency` configuration to alloy-metrics with a default set to 15m.
 - Add grafanaOrganization CRD in helm chart.
 - Add logic to grafanaOrganization reconciler.
+
+### Changed
+
+- Change default default monitoring agent to Alloy
 
 ## [0.7.1] - 2024-10-10
 
@@ -143,7 +155,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.7.1...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.8.1...HEAD
+[0.8.1]: https://github.com/giantswarm/observability-operator/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/giantswarm/observability-operator/compare/v0.7.1...v0.8.0
 [0.7.1]: https://github.com/giantswarm/observability-operator/compare/v0.7.0...v0.7.1
 [0.7.0]: https://github.com/giantswarm/observability-operator/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/giantswarm/observability-operator/compare/v0.6.0...v0.6.1